feat: created advancedResults middleware to import to bootcamps and courses
