+++ controllers/bootcamps.js
@@ -4,75 +4,7 @@ const Bootcamp = require('../models/Bootcamp');
 const asyncHandler = require('../middleware/async');
 
 exports.getBootcamps = asyncHandler(async (req, res, next) => {
-  let query;
-
-  // copy req.query
-  const reqQuery = { ...req.query };
-
-  // fields to exclude
-  const removeFields = ['select', 'sort', 'page', 'limit'];
-
-  // Loop over removeFields and delete from reqQuery
-  removeFields.forEach(param => delete reqQuery[param]);
-
-  // copy query string
-  let queryStr = JSON.stringify(reqQuery);
-
-  // create Mongoose operators ($gt, $gte, etc)
-  queryStr = queryStr.replace(/\b(gt|gte|lt|lte|in)\b/g, match => `$${match}`);
-
-  // finding resource
-  query = Bootcamp.find(JSON.parse(queryStr)).populate('courses');
-
-  //select params
-  if (req.query.select) {
-    const fields = req.query.select.split(',').join(' ');
-    query = query.select(fields);
-  }
-
-  //sort params
-  if (req.query.sort) {
-    const sortBy = req.query.sort.split(',').join(' ');
-    query = query.sort(sortBy);
-  } else {
-    query = query.sort('-createdAt');
-  }
-
-  // pagination and limit
-  const page = parseInt(req.query.page, 10) || 1;
-  const limit = parseInt(req.query.limit, 10) || 25;
-  const startIndex = (page - 1) * limit;
-  const endIndex = page * limit;
-  const total = await Bootcamp.countDocuments();
-
-  query = query.skip(startIndex).limit(limit);
-
-  // Executing query
-  const bootcamps = await query;
-
-  // pagination results
-  const pagination = {};
-
-  if (endIndex < total) {
-    pagination.next = {
-      page: page + 1,
-      limit
-    };
-  }
-
-  if (startIndex > 0) {
-    pagination.prev = {
-      page: page - 1,
-      limit
-    };
-  }
-
-  res.status(200).json({
-    success: true,
-    count: bootcamps.length,
-    pagination,
-    data: bootcamps
-  });
+  res.status(200).json(res.advancedResults);
 });
 
 exports.getBootcamp = asyncHandler(async (req, res, next) => {
