+++ routes/courses.js
@@ -8,6 +8,7 @@ const {
 } = require('../controllers/courses');
 
 const Course = require('../models/Course');
+const advancedResults = require('../middleware/advancedResults');
 
 const router = express.Router({ mergeParams: true });
 
@@ -15,7 +16,11 @@ const { protect } = require('../middleware/auth');
 
 router
   .route('/')
-  .get(protect, getCourses)
+  .get(
+    protect,
+    advancedResults(Course, { path: 'bootcamp', select: 'name description' }),
+    getCourses
+  )
   .post(protect, addCourse);
 
 router
